+++ features/search/components/chip/category.tsx
@@ -4,7 +4,7 @@ import { useMemo, useState } from "react";
 
 import useCategories, { CategoryGroup } from "@/hooks/use-categories";
 import { matcher } from "@/features/search/utils/matcher";
-import { Category } from "@/utils/category";
+import { Category } from "@/utils/paper/category";
 import CategoryIcon from "@/ui/icons/category";
 import ArrowDownIcon from "@/ui/icons/arrow-down";
 import SearchField from "@/ui/search-field";
