+++ hooks/use-journals.ts
@@ -1,5 +1,5 @@
 import { useMemo } from "react";
-import { Journal } from "@/utils/journal";
+import { Journal } from "@/utils/paper/journal";
 
 // TODO: 이후 서버와 연동하는 형태로 구현할 예정입니다.
 export default function useJournals() {
